+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- Replace **letter** with **paper**.
 - README now reflects IEEE Access published the paper.
 - COVID dataset boxplot creation now reflects it in its naming.
 - Montgomery's transformation script now adds an extra dimension at the end of each image, so the processing script can work with them.